requirements.txt
Package                Version
---------------------- -------
appdirs                1.4.3  
attrs                  19.1.0 
black                  19.3b0 
cffi                   1.12.3 
Click                  7.0    
configparser           4.0.2  
Django                 2.2.5  
django-braces          1.13.0 
django-crispy-forms    1.7.2  
django-widget-tweaks   1.4.5  
easy-thumbnails        2.6    
misaka                 2.1.1  
mod-wsgi               4.6.7  
mysql-connector-python 8.0.17 
Pillow                 6.1.0  
pip                    19.2.3 
protobuf               3.9.1  
psycopg2-binary        2.8.3  
pycparser              2.19   
pytz                   2019.2 
setuptools             41.2.0 
six                    1.12.0 
sqlparse               0.3.0  
toml                   0.10.0 
wheel                  0.33.6 
